Refactor grammar into module and remove all mod.rs
